requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.5
distro==1.9.0
dj-rest-auth==7.1.1
Django==6.0.2
django-allauth==65.14.3
django-cors-headers==4.9.0
django-rest-auth==0.9.5
django_celery_results==2.6.0
djangorestframework==3.16.1
dnspython==2.8.0
feedparser==6.0.12
frozenlist==1.8.0
greenlet==3.4.0
groq==1.0.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jwt==1.4.0
kombu==5.6.2
lxml==6.0.2
multidict==6.7.1
packaging==26.0
pillow==12.1.1
playwright==1.58.0
prompt_toolkit==3.0.52
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dateutil==2.9.0.post0
redis==7.2.1
reportlab==4.4.10
requests==2.32.5
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
yarl==1.22.0
